Rename AccountEditor
命名を変えて分かりやすく
